refactor(transform): move api methods and classes to separate files (#138)
