+++ pyproject.toml
@@ -51,6 +51,7 @@ strict = true
 
 mypy_path = ["src"]
 explicit_package_bases = true
+namespace_packages = true
 
 files = [
     "src",
